+++ server/server.go
@@ -12,8 +12,8 @@ import (
 	"github.com/go-redis/redis/v8"
 	"github.com/mikeydub/go-gallery/eth"
 	"github.com/mikeydub/go-gallery/memstore"
-	"github.com/mikeydub/go-gallery/mongodb"
 	"github.com/mikeydub/go-gallery/persist"
+	"github.com/mikeydub/go-gallery/persist/mongodb"
 	"github.com/mikeydub/go-gallery/util"
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
